python/requirements.txt
grpcio==1.76.0
grpcio-tools==1.76.0
protobuf==6.33.2
extremitypathfinder==2.7.2
